Fix location handling in integration test and remove unnecessary classes.
